+++ src/components/Login.js
@@ -62,17 +62,15 @@ const forgotPassLink = "https://www.reddit.com/password";
 
 const LoginPopup = ({ setHide }) => {
 	// used to focus on username input field and to close popup on outside click.
-	const [popup, userInput] = [useRef(), useRef()];
+	const popup = useRef();
 
 	useHotkeys("Escape", () => {
-		userInput.current.blur();
 		setHide(true);
 	});
 
 	const watchEsc = (e) => (e.key === "Escape" ? setHide(true) : null);
 
 	useEffect(() => {
-		userInput.current.focus(); // focusing on the form
 		const watch = (e) => {
 			// we need to check path otherwise clickin a child also hides the popup.
 			!e.path.includes(popup.current) && setHide(true);
@@ -113,12 +111,13 @@ const LoginPopup = ({ setHide }) => {
 			<div className="contain">
 				<h2>Sign In</h2>
 				<input
+					// what are autoCorrect, autoComplete, autoSave
+					autoFocus
 					type="text"
 					className="username"
 					name="username"
 					placeholder="username"
 					onKeyDown={watchEsc}
-					ref={userInput}
 					onChange={onChange}
 					value={username}
 				/>{" "}